+++ platforms/kotlin/src/main/kotlin/ai/snips/nlu/ontology/Ontology.kt
@@ -47,6 +47,7 @@ data class Range @ParcelConstructor constructor(@ParcelProperty("start") val sta
 @Parcel(BEAN)
 data class Slot @ParcelConstructor constructor(@ParcelProperty("rawValue") val rawValue: String,
                                                @ParcelProperty("value") val value: SlotValue,
+                                               @ParcelProperty("alternatives") val alternatives: List<SlotValue>,
                                                @ParcelProperty("range") val range: Range,
                                                @ParcelProperty("entity") val entity: String,
                                                @ParcelProperty("slotName") val slotName: String,
@@ -171,7 +172,13 @@ data class IntentClassifierResult @ParcelConstructor constructor(
         @ParcelProperty("confidenceScore") val confidenceScore: Float)
 
 @Parcel(BEAN)
+data class IntentParserAlternative @ParcelConstructor constructor(
+        @ParcelProperty("intent") val intent: IntentClassifierResult,
+        @ParcelProperty("slots") val slots: List<Slot>)
+
+@Parcel(BEAN)
 data class IntentParserResult @ParcelConstructor constructor(
         @ParcelProperty("input") val input: String,
         @ParcelProperty("intent") val intent: IntentClassifierResult,
-        @ParcelProperty("slots") val slots: List<Slot>)
+        @ParcelProperty("slots") val slots: List<Slot>,
+        @ParcelProperty("alternatives") val alternatives: List<IntentParserAlternative>)