+++ src/ontology.rs
@@ -5,6 +5,13 @@ pub struct IntentParserResult {
     pub input: String,
     pub intent: IntentClassifierResult,
     pub slots: Vec<Slot>,
+    pub alternatives: Vec<IntentParserAlternative>,
+}
+
+#[derive(Serialize, Deserialize, Debug, Clone, PartialEq)]
+pub struct IntentParserAlternative {
+    pub intent: IntentClassifierResult,
+    pub slots: Vec<Slot>,
 }
 
 #[derive(Serialize, Deserialize, Debug, Clone, PartialEq)]
@@ -19,10 +26,11 @@ pub struct IntentClassifierResult {
 pub struct Slot {
     pub raw_value: String,
     pub value: SlotValue,
+    pub alternatives: Vec<SlotValue>,
     pub range: Range<usize>,
     pub entity: String,
     pub slot_name: String,
-    #[serde(skip_serializing_if = "Option::is_none")] 
+    #[serde(skip_serializing_if = "Option::is_none")]
     pub confidence_score: Option<f32>,
 }
 
@@ -33,10 +41,12 @@ impl Slot {
         entity: String,
         slot_name: String,
         confidence_score: Option<f32>,
+        alternatives: Vec<SlotValue>,
     ) -> Slot {
         Slot {
             raw_value: value.clone(),
             value: SlotValue::Custom(value.into()),
+            alternatives,
             range,
             entity,
             slot_name,
@@ -50,6 +60,7 @@ impl Slot {
         Slot {
             raw_value: self.raw_value,
             value: slot_value,
+            alternatives: self.alternatives,
             range: self.range,
             entity: self.entity,
             slot_name: self.slot_name,
@@ -181,6 +192,7 @@ mod tests {
         let slot = Slot {
             raw_value: "value".into(),
             value: SlotValue::Custom("value".into()),
+            alternatives: vec![],
             range: 0..5,
             entity: "toto".into(),
             slot_name: "toto".into(),
@@ -195,6 +207,7 @@ mod tests {
         let slot = Slot {
             raw_value: "fifth".into(),
             value: SlotValue::Ordinal(OrdinalValue { value: 5 }),
+            alternatives: vec![],
             range: 0..5,
             entity: "toto".into(),
             slot_name: "toto".into(),
@@ -213,6 +226,7 @@ mod tests {
                 grain: Grain::Year,
                 precision: Precision::Exact,
             }),
+            alternatives: vec![],
             range: 0..10,
             entity: "toto".into(),
             slot_name: "toto".into(),