+++ CHANGELOG.md
@@ -1,13 +1,23 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## [0.66.0]
+## [Unreleased]
+### Added
+- Add an `alternatives` attribute, containing less likely alternatives, to ([#153](https://github.com/snipsco/snips-nlu-ontology/pull/153)):
+    - `IntentParserResult`
+    - `Slot`
+    - `BuiltinEntity`
+
+### Fixed
+- Update kotlin wrapper from 4.1 to 5.1
+
+## [0.66.0] - 2019-07-17
 ### Changed
 - Remove references to entity parsing [#151](https://github.com/snipsco/snips-nlu-ontology/pull/151)
 - Remove support for music entities in korean [#148](https://github.com/snipsco/snips-nlu-ontology/pull/148)
 - Remove snips-nlu-ontology-export kotlin package [#149](https://github.com/snipsco/snips-nlu-ontology/pull/149)
 
-## [0.65.0]
+## [0.65.0] - 2019-07-11
 ### Added
 - Add datetime subtypes [#144](https://github.com/snipsco/snips-nlu-ontology/pull/144)
 - Add geographical entities [#145](https://github.com/snipsco/snips-nlu-ontology/pull/145)
@@ -191,6 +201,7 @@ All notable changes to this project will be documented in this file.
 ### Changed
 - Updated Rustling ontology to `0.16.4`
 
+[Unreleased]: https://github.com/snipsco/snips-nlu-ontology/compare/0.66.0...HEAD
 [0.66.0]: https://github.com/snipsco/snips-nlu-ontology/compare/0.65.0...0.66.0
 [0.65.0]: https://github.com/snipsco/snips-nlu-ontology/compare/0.64.8...0.65.0
 [0.64.8]: https://github.com/snipsco/snips-nlu-ontology/compare/0.64.7...0.64.8